src/webUI/requirements.txt
flask
flask-cors
firebase
firebase-admin
python-dotenv